requirements.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
cmake==4.0.2
diskcache==5.6.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
numpy==2.3.0
python-dotenv==1.1.0
python-telegram-bot==22.1
requests==2.32.4
sniffio==1.3.1
typing_extensions==4.14.0
urllib3==2.4.0